moved TestIsRedemptionRateWithinSafetyBounds
